requirements.txt
Click
dash
dash-core-components
dash-html-components
dash-renderer
Flask
Flask-Compress
future
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
pytz
retrying
six
Werkzeug
gunicorn
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-resumable-media
googleapis-common-protos
google-cloud-bigquery
db-dtypes
prophet